fine_tuning/requirements.txt
torch>=2.2.0
transformers>=4.36.2
datasets>=2.16.1
peft>=0.7.1
bitsandbytes>=0.41.3
accelerate>=0.25.0
wandb>=0.16.2
huggingface-hub>=0.20.3
python-dotenv>=1.0.0
scipy>=1.11.4
sentencepiece>=0.1.99
protobuf>=4.25.2
paramiko>=3.0.0
requests>=2.28.0 